% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit_varQ.R
\name{crit_varQ}
\alias{crit_varQ}
\title{Future quantile variance}
\usage{
crit_varQ(xnew, model, data.x, x, alpha, beta = 4)
}
\arguments{
\item{xnew}{the new point}

\item{model}{object of class km}

\item{data.x}{data related to x}

\item{x}{integration points}

\item{alpha}{quantile level}

\item{beta}{a threshold}
}
\value{
the criterion
}
\examples{
\dontrun{
compute_crit(xnew, model, data.x, x, alpha, beta=4)
}
}
\references{
T. Labopin-Richard, V. Picheny, "Sequential design of experiments for estimating quantiles of black-box functions", 
Statistica Sinica, 2017, \emph{doi:10.5705/ss.202016.0160}
}
\author{
Victor Picheny
}
